backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
Flask-JWT-Extended==4.4.4
PyMySQL==1.0.3
marshmallow==3.19.0
python-dotenv==1.0.0
Werkzeug==2.0.1 